# src/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Data API, database migration scripts, and backend services for all Podverse models.",
    "version": "1.0.0",
    "title": "Podverse API",
    "termsOfService": "https://podverse.fm/terms/",
    "contact": {
      "email": "dev@podverse.fm"
    },
    "license": {
      "name": "GNU Affero General Public License v3",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
  "host": "localhost:1234",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login as user",
        "description": "Sets a bearer token in the Authentication cookie",
        "operationId": "authLogin",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "email": "string",
                "password": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset a user's password",
        "description": "Reset a user's password if the correct token is provided",
        "operationId": "authResetPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "password": "string",
                "resetPasswordToken": "guid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "User not found or invalid token"
          }
        }
      }
    },
    "/auth/send-reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send email to reset user's password",
        "description": "Sends email to user that directs the user to the password reset page, along with a resetPasswordToken",
        "operationId": "authSendResetPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "email": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized, token has expired"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/auth/sign-up": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign up new user",
        "description": "Create a new user, and send verification email",
        "operationId": "authSignUp",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "email": "string",
                "name": "string",
                "password": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/send-verification": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send email to verify a user's email address",
        "description": "Send email to verify a user's email address",
        "operationId": "authSendVerification",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "email": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/auth/verify-email": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Verify a user's email address",
        "description": "Verification is successful if a valid emailVerificationToken is included in the request",
        "operationId": "authVerifyEmail",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized, token has expired"
          },
          "404": {
            "description": "User not found or invalid verification token"
          }
        }
      }
    },
    "/author": {
      "get": {
        "tags": [
          "author"
        ],
        "summary": "Find authors by fields",
        "description": "Returns an array of authors",
        "operationId": "getAuthors",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "slug",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Author"
            }
          }
        }
      }
    },
    "/author/{authorId}": {
      "get": {
        "tags": [
          "author"
        ],
        "summary": "Find author by ID",
        "description": "Returns a single author",
        "operationId": "getAuthorById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "description": "ID of author to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Author"
            }
          },
          "404": {
            "description": "Author not found"
          }
        }
      }
    },
    "/category": {
      "get": {
        "tags": [
          "category"
        ],
        "summary": "Find categories by fields",
        "description": "Returns an array of categories",
        "operationId": "getCategories",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "slug",
            "in": "query",
            "type": "string"
          },
          {
            "name": "title",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
    "/category/{categoryId}": {
      "get": {
        "tags": [
          "category"
        ],
        "summary": "Find category by ID",
        "description": "Returns a single category",
        "operationId": "getCategoryById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "ID of category to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
    "/episode": {
      "get": {
        "tags": [
          "episode"
        ],
        "summary": "Find episodes by fields",
        "description": "Returns an array of episodes",
        "operationId": "getEpisodes",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authors",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "categories",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "type": "string"
          },
          {
            "name": "duration",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "episodeType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "guid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "imageUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isExplicit",
            "in": "query",
            "type": "boolean"
          },
          {
            "name": "linkUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "mediaUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "mediaRefs",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "podcast",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pubDate",
            "in": "query",
            "type": "string",
            "format": "date"
          },
          {
            "name": "title",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Episode"
            }
          }
        }
      }
    },
    "/episode/{episodeId}": {
      "get": {
        "tags": [
          "episode"
        ],
        "summary": "Find episode by ID",
        "description": "Returns a single episode",
        "operationId": "getEpisodeById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "episodeId",
            "in": "path",
            "description": "ID of episode to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Episode"
            }
          },
          "404": {
            "description": "Episode not found"
          }
        }
      }
    },
    "/feedUrl": {
      "get": {
        "tags": [
          "feedUrl"
        ],
        "summary": "Find feedUrls by fields",
        "description": "Returns an array of feedUrls",
        "operationId": "getFeedUrls",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "isAuthority",
            "in": "query",
            "type": "boolean"
          },
          {
            "name": "podcast",
            "in": "query",
            "type": "string"
          },
          {
            "name": "url",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/FeedUrl"
            }
          }
        }
      }
    },
    "/feedUrl/{feedUrlId}": {
      "get": {
        "tags": [
          "feedUrl"
        ],
        "summary": "Find feedUrl by ID",
        "description": "Returns a single feedUrl",
        "operationId": "getFeedUrlById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "feedUrlId",
            "in": "path",
            "description": "ID of feedUrl to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/FeedUrl"
            }
          },
          "404": {
            "description": "FeedUrl not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "authorization": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Author": {
      "type": "object",
      "required": [
        "id",
        "name",
        "slug"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        }
      }
    },
    "Category": {
      "type": "object",
      "required": [
        "id",
        "slug",
        "title"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        }
      }
    },
    "Episode": {
      "type": "object",
      "required": [
        "id",
        "mediaUrl"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "mediaUrl": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "duration": {
          "type": "integer"
        },
        "episodeType": {
          "type": "string"
        },
        "guid": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "isExplicit": {
          "type": "boolean"
        },
        "isPublic": {
          "type": "boolean"
        },
        "linkUrl": {
          "type": "string"
        },
        "mediaFilesize": {
          "type": "integer"
        },
        "mediaType": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "pubDate": {
          "type": "string",
          "format": "date"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "mediaRefs": {
          "$ref": "#/definitions/MediaRef"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "FeedUrl": {
      "type": "object",
      "required": [
        "url",
        "podcast"
      ],
      "properties": {
        "url": {
          "type": "string"
        },
        "isAuthority": {
          "type": "boolean"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "MediaRef": {
      "type": "object",
      "required": [
        "id",
        "episodeMediaUrl",
        "startTime"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "episodeMediaUrl": {
          "type": "string"
        },
        "startTime": {
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "endTime": {
          "type": "integer"
        },
        "episodeDuration": {
          "type": "integer"
        },
        "episodeGuid": {
          "type": "string"
        },
        "episodeId": {
          "type": "string"
        },
        "episodeImageUrl": {
          "type": "string"
        },
        "episodeLinkUrl": {
          "type": "string"
        },
        "episodePubDate": {
          "type": "string",
          "format": "date"
        },
        "episodeSummary": {
          "type": "string"
        },
        "episodeTitle": {
          "type": "string"
        },
        "isPublic": {
          "type": "boolean"
        },
        "ownerId": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "podcastFeedUrl": {
          "type": "string"
        },
        "podcastGuid": {
          "type": "string"
        },
        "podcastId": {
          "type": "string"
        },
        "podcastImageUrl": {
          "type": "string"
        },
        "podcastIsExplicit": {
          "type": "boolean"
        },
        "podcastTitle": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "episode": {
          "$ref": "#/definitions/Episode"
        },
        "owner": {
          "$ref": "#/definitions/User"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "Playlist": {
      "type": "object",
      "required": [
        "id",
        "owner"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "isMyClips": {
          "type": "boolean"
        },
        "isPublic": {
          "type": "boolean"
        },
        "itemsOrder": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "mediaRefs": {
          "$ref": "#/definitions/MediaRef"
        },
        "owner": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "Podcast": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "feedLastUpdated": {
          "type": "string",
          "format": "date"
        },
        "guid": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "isExplicit": {
          "type": "boolean"
        },
        "language": {
          "type": "string"
        },
        "linkUrl": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "episodes": {
          "$ref": "#/definitions/Episode"
        },
        "feedUrls": {
          "$ref": "#/definitions/FeedUrl"
        }
      }
    },
    "User": {
      "type": "object",
      "required": [
        "id",
        "email"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "subscribedPodcastIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "playlists": {
          "$ref": "#/definitions/Playlist"
        }
      }
    }
  }
}